Elimino fetchs innecesarios en componente raiz
